requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
sqlalchemy>=2.0.0
openai>=1.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
reportlab>=4.1.0
PyJWT>=2.8.0
pytest>=8.0.0
# Postgres driver for Cloud SQL. Pure-Python = no compilation needed in the
# slim container. SQLAlchemy talks to it via the postgresql+pg8000 dialect.
pg8000>=1.30.0
